cleaned up, added comments and null checks, and gradle update
