+++ Java/simple-signing-demo/gradle/gradle-daemon-jvm.properties
@@ -1,0 +1,2 @@
+#This file is generated by updateDaemonJvm
+toolchainVersion=21
