+++ Java/simple-signing-demo/src/main/java/co/upvest/client/SignatureComponentsAuth.java
@@ -17,56 +17,35 @@
 package co.upvest.client;
 
 
-import org.slf4j.Logger;
-
 import java.net.URI;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
-import java.security.SignatureException;
-import java.util.LinkedHashMap;
 import java.util.Objects;
 
-public class SignatureComponentsAuth extends SignatureComponents {
-    public final static String httpMethod = "POST";
-    private final static Logger logger = org.slf4j.LoggerFactory.getLogger(SignatureComponentsAuth.class);
-    public final String bodyContent;
-    public final String contentType;
+/**
+ * Container for the mandatory components of the signature for the Auth endpoint.
+ * @param url
+ * @param accept
+ * @param apiVersion
+ * @param bodyContent
+ * @param contentType
+ */
+public record SignatureComponentsAuth(URI url, String accept, String apiVersion, String bodyContent,
+                                      String contentType) {
 
-    public SignatureComponentsAuth(URI url, String accept, String apiVersion, String bodyContent, String contentType) {
-        super(url, accept, apiVersion);
-        Objects.requireNonNull(bodyContent, "bodyContent must not be null");
-        Objects.requireNonNull(contentType, "contentType must not be null");
-        this.bodyContent = bodyContent;
-        this.contentType = contentType;
-    }
+    public SignatureComponentsAuth {
+        Objects.requireNonNull(url, "url cannot be null");
+        Objects.requireNonNull(accept, "accept cannot be null");
+        Objects.requireNonNull(apiVersion, "apiVersion cannot be null");
+        Objects.requireNonNull(bodyContent, "bodyContent cannot be null");
+        Objects.requireNonNull(contentType, "contentType cannot be null");
 
-
-    public SignatureAuth sign(SignatureKey key) {
-        try {
-            var contentDigest = SigningUtil.getSHA512Digest(bodyContent);
-
-            // It has to be a LinkedHashMap to keep the order of the fields
-            var signatureFields = new LinkedHashMap<String, String>();
-            signatureFields.put("\"content-type\"", contentType);
-            signatureFields.put("\"content-length\"", String.valueOf(bodyContent.length()));
-            signatureFields.put("\"accept\"", accept);
-            signatureFields.put("\"@method\"", httpMethod);
-            signatureFields.put("\"@path\"", url.getPath());
-            signatureFields.put("\"content-digest\"", contentDigest);
-
-            var signatureParams = SigningUtil.createSignatureParametersString(signatureFields, key);
-            signatureFields.put("\"@signature-params\"", signatureParams);
-
-            var signature = SigningUtil.sign(signatureFields, key);
-            logger.debug("Signature: {}", signature);
-            return new SignatureAuth(contentDigest, signatureParams, signature);
-        } catch (NoSuchAlgorithmException | InvalidKeyException | SignatureException ex) {
-            logger.error(ex.getMessage(), ex);
-            throw new RuntimeException(ex);
+        if (accept.isEmpty() || apiVersion.isEmpty() || bodyContent.isEmpty() || contentType.isEmpty()) {
+            throw new IllegalArgumentException("String arguments cannot be empty");
         }
     }
 
-    public record SignatureAuth(String contentDigest, String signatureParams, String signature) {
+    public String httpMethod() {
+        return "POST";
     }
-
 }
+
+
