+++ Java/simple-signing-demo/src/main/java/co/upvest/client/SignatureGet.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright © 2024 Upvest GmbH <support@upvest.co>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package co.upvest.client;
+
+import java.util.Objects;
+
+public record SignatureGet(String signatureParams, String signature) {
+    public SignatureGet {
+        Objects.requireNonNull(signatureParams, "signatureParams cannot be null");
+        Objects.requireNonNull(signature, "signature cannot be null");
+
+        if (signatureParams.isEmpty() || signature.isEmpty()) {
+            throw new IllegalArgumentException("String arguments cannot be empty");
+        }
+    }
+}
